Eslint automatic fixes
