+++ front/.eslintignore
@@ -1,0 +1,2 @@
+src/serviceWorker.js
+src/App.test.js
