--- front/src/components/SourceCode/GrammarSourceCode.js
@@ -1,13 +1,0 @@
-import { SourceCode } from "./SourceCode";
-import { connect } from "react-redux";
-
-function MapStateToProps (state) {
-    return ({
-        style: !state ? {background: '#FFFFFF'} : {background: '#A9A9A9'},
-        readOnly: state,
-        highlightActiveLine: !state,
-        value: localStorage.getItem("syntax")
-    })
-}
-
-export default connect(MapStateToProps) (SourceCode)
