+++ front/src/components/SourceCode/GrammarSourceCode.jsx
@@ -3,10 +3,10 @@ import { SourceCode } from './SourceCode';
 
 function MapStateToProps(state) {
   return ({
-    style: state ? { background: '#FFFFFF' } : { background: '#A9A9A9' },
-    readOnly: !state,
-    highlightActiveLine: state,
-    value: localStorage.getItem('sourceCode'),
+    style: !state ? { background: '#FFFFFF' } : { background: '#A9A9A9' },
+    readOnly: state,
+    highlightActiveLine: !state,
+    value: localStorage.getItem('syntax'),
   });
 }
