+++ front/src/components/SourceCode/MapStateToProps.js
@@ -1,7 +1,5 @@
-export default (state) => {
-    return ({
-        style: state ? {background: '#FFFFFF'} : {background: '#A9A9A9'},
-        readOnly: !state,
-        highlightActiveLine: state
-    })
-}
+export default (state) => ({
+  style: state ? { background: '#FFFFFF' } : { background: '#A9A9A9' },
+  readOnly: !state,
+  highlightActiveLine: state,
+});
