--- front/src/components/SourceCode/index.js
@@ -1,2 +1,0 @@
-export * from './SourceCode.js'
-export * from './ProgramSourceCode.js'
