+++ front/src/components/SourceCode/index.jsx
@@ -1,0 +1,2 @@
+export * from './SourceCode.js';
+export * from './ProgramSourceCode.js';
