+++ front/src/index.jsx
@@ -3,19 +3,19 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-import { syntaxExample, semanticsExample, sourceCodeExample } from './example'
-import './assets/scss/auth.scss'
-import './App.css'
+import { syntaxExample, semanticsExample, sourceCodeExample } from './example';
+import './assets/scss/auth.scss';
+import './App.css';
 
-localStorage.setItem("sourceCode", sourceCodeExample)
-localStorage.setItem("syntax", syntaxExample)
-localStorage.setItem("semantics", semanticsExample)
+localStorage.setItem('sourceCode', sourceCodeExample);
+localStorage.setItem('syntax', syntaxExample);
+localStorage.setItem('semantics', semanticsExample);
 
 ReactDOM.render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
-  document.getElementById('root')
+  document.getElementById('root'),
 );
 
 // If you want your app to work offline and load faster, you can change
