+++ front/src/reducers/SourceCodeFieldReducer.js
@@ -1,11 +1,11 @@
 const initialState = false;
 
 export function SourceCodeFieldReducer(state = initialState, action) {
-  console.log('Changing state: ' + action.type);
-  switch(action.type) {
-    case "WRITING_PROGRAM":
+  console.log(`Changing state: ${action.type}`);
+  switch (action.type) {
+    case 'WRITING_PROGRAM':
       return true;
-    case "WRITING_GRAMMAR":
+    case 'WRITING_GRAMMAR':
       return false;
     default:
       return state;
