+++ front/src/utilities/Queries.js
@@ -1,37 +1,39 @@
 function getCookieByName(name) {
-    var match = document.cookie.match(new RegExp('(^| )' + name + '=([^;]+)'));
-    if (match) return match[2];
+  const match = document.cookie.match(new RegExp(`(^| )${name}=([^;]+)`));
+  return match?.[2];
 }
 
-function deleteCookie( name ) {
-    if( getCookieByName( name ) ) {
-      document.cookie = name + "=" +
+function deleteCookie(name) {
+  if (getCookieByName(name)) {
+    document.cookie = `${name}=`
         // ((path) ? ";path="+path:"")+
         // ((domain)?";domain="+domain:"") +
-        ";expires=Thu, 01 Jan 1970 00:00:01 GMT";
-    }
+        + ';expires=Thu, 01 Jan 1970 00:00:01 GMT';
   }
+}
 
 const fetchGet = async (url) => fetch(url, {
-    method: 'GET',
-    headers: {
-        'Content-Type': 'application/json',
-        'Accept': 'application/json',
-        'username': getCookieByName('username'),
-    }
+  method: 'GET',
+  headers: {
+    'Content-Type': 'application/json',
+    Accept: 'application/json',
+    username: getCookieByName('username'),
+  },
 });
 
 const fetchPostJson = async (url, json) => fetch(url, {
-    method: 'POST',
-    headers: {
-        'Content-Type': 'application/json',
-        'Accept': 'application/json',
-        'username': getCookieByName('username'),
-        'project': getCookieByName('project')
-    },
-    body: json
+  method: 'POST',
+  headers: {
+    'Content-Type': 'application/json',
+    Accept: 'application/json',
+    username: getCookieByName('username'),
+    project: getCookieByName('project'),
+  },
+  body: json,
 });
 
-export { fetchGet, fetchPostJson, getCookieByName, deleteCookie }
+export {
+  fetchGet, fetchPostJson, getCookieByName, deleteCookie,
+};
 
-export default fetchPostJson
+export default fetchPostJson;
